Beginnings of Quest Tracking. Created QuestTracker (stored in GameControl), Quest and QuestData classes for storing and tracking individual quest information. Modified NPC dialogue to handle saying different dialogues depending on any active quests. (Also fixed bug with DialogueResponse menu not showing all buttons)
More testing and modifications needed. Such as reward system implementation. Also UI notifications -- when a quest is started, when a  subquest is completed and new task starts, and when the whole quest is completed.
Also will eventually need descriptions for each quest that will be shown at some part in the pause menu -- show name of quest, a summary of the whole quest, (maybe what subquests within the quest the player has already completed), and what the current goal of the quest is